+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All significant changes to this project will be documented in this file.
 - `GenericCoreDataStack` now has a generic parameter for `ViewContextType` which must be specified.
 - Changed `GenericCoreDataStack` & `CoreDataStack` `viewContext` to use the `ViewContextType` removing the read only context.
 - Changed default merge policy for `ActionContext` to `NSMergeByPropertyObjectTrumpMergePolicy`
+- Changed `NSEntityDescription.uniquenessAttributes` to be non-optional.
 
 ## [3.0.0-beta.2](https://github.com/tonystone/coherence/releases/tag/3.0.0-beta.2)